python/examples/openai/requirements.txt
lightspark
python-dotenv
openai
